Move files from submit-queue to mungegithub
