+++ mungegithub/committers.txt
