--- submit-queue/.gitignore
@@ -1,1 +1,0 @@
-submit-queue
