--- submit-queue/pod.yaml
@@ -1,38 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: submit-queue
-  namespace: default
-  labels:
-    app: submit-queue
-    version: "0.12"
-spec:
-  containers:
-  - image: gcr.io/google_containers/submit-queue:0.12
-    command:
-    - ./submit-queue
-    - --jenkins-host=http://jenkins-master:8080
-    - --token-file=/etc/secret-volume/token
-    - --jenkins-jobs=kubernetes-e2e-gce,kubernetes-e2e-gke-ci,kubernetes-build,kubernetes-e2e-gce-parallel,kubernetes-e2e-gce-reboot,kubernetes-e2e-gce-scalability
-    - --alsologtostderr
-    - --v=5
-    - --www=/www
-    imagePullPolicy: IfNotPresent
-    name: submit-queue
-    volumeMounts:
-    - name: secret-volume
-      mountPath: /etc/secret-volume
-    ports:
-    - name: status
-      containerPort: 8080
-    livenessProbe:
-      httpGet:
-        path: /api
-        port: 8080
-      initialDelaySeconds: 10
-      timeoutSeconds: 1
-  volumes:
-    - name: secret-volume
-      secret:
-        secretName: github-token
-  restartPolicy: OnFailure
